pyproject.toml
[tool.poetry]
name = "fastapi_playground"
version = "0.1.0"
description = ""
authors = ["Leung Wai Lun <williamleung@atal.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.85.0"
loguru = "^0.6.0"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
pendulum = "^2.1.2"
aiohttp = "^3.8.3"
requests = "^2.28.1"
python-multipart = "^0.0.5"
python-keycloak = "^2.5.0"
pandas = "^1.5.1"
paho-mqtt = "^1.6.1"
kafka = "^1.3.5"
confluent-kafka = "^1.9.2"
boto3 = "^1.26.32"
pydantic = {extras = ["email"], version = "^1.10.2"}
docker = "^6.0.1"
asyncio-mqtt = "^0.16.2"

[tool.poetry.dev-dependencies]
black = "^22.8.0"
pylint = "^2.15.3"
mypy = "^0.981"
autoflake = "^1.6.1"
pytest = "^7.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]] 
name = "internal" 
url = "http://pip.atal.energy/" 
secondary = true